Sync up FFmpeg -- autoconf
This changeset is the first in a series to sync up our FFmpeg code with a
much more modern and up to date FFmpeg.  This one takes care of the
autoconf side of things.  It is cherry-picked from an earlier ffmpeg-fix
branch, squashing all the configuration related changes.

This changeset also introduces new minimum versions of libavformat and
libavcodec required by SXEmacs:

  libavformat -- 55.12.100
  libavcodec  -- 55.18.102

	* configure.ac: Need libs for avcodec as well as avformat. Add
          in libavutil libs.

	* m4/sxe-mm.m4 (SXE_MM_CHECK_FFMPEG): Rework it for modern
	ffmpeg. Check for av_dict_get() in libavutil.

	* m4/sxe-mm.m4 (SXE_CHECK_FFMPEG_HEADERS): Check for avcodec.h
	and dict.h as well.

	* m4/sxe-mm.m4 (SXE_CHECK_FFMPEG_LIBS): Check for libavutil.

	* m4/sxe-mm.m4 (SXE_MM_CHECK_FFMPEG):