Re-write the git-for-steve.sh script
Many thanks and kudos to Nelson for not only writing the original script,
but for putting up with my endless questions on IRC.

The old script was great, and in fact I referenced it countless times in
the past to find out how to do certain git things, but it only went as far
as to print suggestions to stdout as to what you should change.  This new
improved incarnation goes much further.

Here is what it _won't_ do:

  It won't do anything if you're not in a SXEmacs source tree
  It won't change anything that is already set without asking
  It won't mess up a dirty WD, so you could run it right in the middle of
	hacking if you wanted to
  It won't do the wrong thing if you run it a 2nd, 3rd, 4th time (all bets
	are off for the 5th run, no I'm kidding, it is safe to run as
	often as you like)

What it does do is take care of everything SXEmacs-git related that you
can think of and probably a few things that you haven't.

	* contrib/git-for-steve.sh: Complete re-write.

Signed-off-by: Steve Youngs <steve@sxemacs.org>