Fix bugs in git-for-steve.sh script,
This changeset fixes one of those "hard to find" errors you get in shell
scripts that turn out to be a single character fix.  Sheesh.  So in the 2
days I spent tracking down this bloody 1 char fix I also tidied up the
rest of the script and improved the UI a tad.  Also fixed a couple of
other bugs that hadn't been discovered yet. :)

	* contrib/git-for-steve.sh: Fix/improve quoting throughout.
	Some UI improvements.
	(set_aliases): Add missing ')'.
	(set_keys): Grep for '^default-key' instead of 'default-key' in
	gpg.conf in case it is commented out.

Signed-off-by: Steve Youngs <steve@sxemacs.org>