Define `pi' in elisp, even with bigfr. (Closes bug #176)
An elisp constant defined with #'defconst can be let-bound, or otherwise
altered.  But one that is defined in C cannot be.  SXEmacs' pi is
sometimes defined in elisp, and sometimes in C, depending on whether it
has bigfr or not.  The Calc XEmacs package let-binds pi in a couple of
places, resulting in errors when (featurep 'bigfr) => t.

With this change, pi is defined in elisp regardless of bigfr.  But it will
still use bigfr to obtain the value if it is available.

	* lisp/float-sup.el: Define pi in elisp, even if we have bigfr,
	and in that case set it to what mpfr gives us.
	Same for e.

	* src/ent/ent-mpfr.c (init_ent_mpfr): Vpi -> Vbigfr_pi
	(vars_of_ent_mpfr): Vpi -> Vbigfr_pi, pi -> bigfr-pi

	* src/ent/ent-mpfr.h: Vpi -> Vbigfr_pi

Signed-off-by: Steve Youngs <steve@sxemacs.org>