Prevent Unexpected errors in OpenSSL Testsuite.
For some years now we've been seeing a couple of unexpected errors
cropping up in the OpenSSL tests.  It turns out that certain ciphers
disagree with how we do things, or perhaps OpenSSL itself is buggy.
Either way, what this changeset does is stop those ciphers being used in
the testsuite.

Yes, this is very much a case of sticking our fingers in our ears and
screaming "LA LA LA I'm NOT LISTENING!!".  Expect a real fix in openssl.c
to come in the future.  Hopefully sooner, rather than later.

	* tests/automated/openssl-tests.el: Prevent known-bad ciphers from
	ruining our testsuite.
	Open up the tests to also include ciphers smaller than 128bits,
	and all digests. This effectively increases the number of tests
	run to more than 200,000.

Signed-off-by: Steve Youngs <steve@sxemacs.org>