Automake/Libtool warning fix/suppression
This changeset cleans up all of the warnings that our autogen.sh spits
out.  At least on my system with automake 1.15, and libtool 2.4.6.

A possibly controversial change here (read: ugly as hell) is that I've
symlinked a couple of files from `src' into `lib-src'.  The
lib-src/Makefile.am referenced regex.c, regex.h, and alloca.c in 'src'.
That caused automake to freak out and warn that we were using source files
in subdirectories which would become incompatible in the future.  Automake
does have a "subdir-objects" option for this type of thing, but it causes
problems elsewhere in our build chain.

The symlink "fix" is quick, dirty, and easy, however I'm happy to have it
reverted if a neater solution can be found.

	* m4/sxe-libtool.m4 (_SXE_CHECK_LT2): Pull in `_LT_COMPILER_PIC'
	to fix a swag of warnings about macros being expanded before being
	required.

	* lib-src/Makefile.am (REGEX_C, ALLOCA_C): Point to files in the
	current directory which are symlinks to the originals in `src'.
	This prevents a scary warning from automake about files in
	subdirectories and future incompatibities.  Automake does have a
	"subdir-objects" option, but because we have a specialised build
	that causes more problems than it fixes.

	* configure.ac (AM_INIT_AUTOMAKE): Use `-Wno-override' and
	`-Wno-portability' to suppress harmless and irrelevant, at least
	to us, warnings.

Signed-off-by: Steve Youngs <steve@sxemacs.org>