Let's not do "Pre-Release" packages.
Having "Pre-Release" packages and "Official Release" packages was a good
idea back in the early part of 2003 when I added that to XEmacs.  There
was an incredible amount of activity happening in the packages in those
days and bugs were frequently slipping into the packages releases.  Today,
things are different.  The amount of activity in packages, and the number
of users/devs we have it should be easily possible to keep everything in
check without the need for a separate "testing" package hierarchy.

We can always reimplement something along the lines of pre-release
packages at a later date if things get out of hand.

	* lisp/menubar-items.el (default-menubar): Remove "Pre-Release"
	package menu.

	* lisp/package-ui.el (package-ui-pre-release-download-menu):
	Remove.

	* lisp/package-get.el (package-get-download-sites): Add our FTP,
	remove everyone else.
	(package-get-pre-release-download-sites): Remove.

Signed-off-by: Steve Youngs <steve@sxemacs.org>