Update alist.el -- Use the newer version from APEL.
Ever since our inception the copy of alist.el in the APEL package has
shadowed the one we have in core.  As the one in APEL is slightly
newer, I am replacing the core alist.el with that version.  The one in
the APEL package will be marked as "suppressed".

	* lisp/alist.el: Updated from the copy of this lib in the APEL
	package.

Signed-off-by: Steve Youngs <steve@sxemacs.org>