Texinfo 6.7 fixes
This makes our docs compatible with texinfo-6.7.  There was one file,
sxemacs-faq.texi, that I ended up converting to UTF-8.  Please be
aware and take care when editing that one from now on.

I did the conversion outside of SXEmacs with iconv and sed, so as to
eliminate any risk of corruption.  Here is how...

,----
| iconv -f ISO-8859-1 -t UTF-8//TRANSLIT -o utf-sxemacs-faq.texi
|   sxemacs-faq.texi &&
|
| sed -i -e 's/iso-2022-8/utf-8/' -e '3i@documentencoding UTF-8' \
|   utf-sxemacs-faq.texi &&
| mv -v {utf-,}sxemacs-faq.texi
`----

I did attempt to find any occurances of chars it couldn't convert, but
I'm sure you'll agree that searching for the odd out of place `?' in
a large FAQ... well, "needle in a hay-stack" comes to mind.

	* info/sxemacs-faq.texi: Convert file t UTF-8.
	Set @documentencoding UTF-8
	(Q1.3.8): Fix a @ref

	* info/sppm.texi (Sending a patch from a git repo): Remove
	@indent from within `@code' sections. It's frowned upon in
	texinfo 6.7.

	* info/lispref/ldap.texi (LDAP Variables): Fix file
        corruption.
        Reported-by: Andrés Ramírez <rrandresf@gmail.com>

Signed-off-by: Steve Youngs <steve@sxemacs.org>